climbcoach_mvp/requirements.txt
opencv-python
mediapipe
numpy
PyYAML
shapely
matplotlib
pyttsx3
websockets
